Синхронизация

+++ code/game/objects/structures/tables_racks.dm
@@ -25,6 +25,9 @@
 	var/framestack = /obj/item/stack/rods
 	var/buildstack = /obj/item/stack/sheet/metal
 	var/busy = 0
+	var/buildstackamount = 1
+	var/framestackamount = 2
+	var/mob/tableclimber
 
 /obj/structure/table/New()
 	..()
@@ -180,26 +183,15 @@
 		else
 			dir = 2
 
-/obj/structure/table/ex_act(severity)
-	switch(severity)
-		if(1.0)
-			qdel(src)
-			return
-		if(2.0)
-			if (prob(50))
-				qdel(src)
-				return
-		if(3.0)
-			if (prob(25))
-				src.density = 0
-		else
-	return
-
+/obj/structure/table/ex_act(severity, target)
+	..()
+	if(severity == 3)
+		if(prob(25))
+			table_destroy(1)
 
 /obj/structure/table/blob_act()
 	if(prob(75))
 		table_destroy(1)
-		qdel(src)
 		return
 
 /obj/structure/table/attack_alien(mob/living/user)
@@ -218,15 +210,20 @@
 /obj/structure/table/attack_paw(mob/user)
 	attack_hand(user)
 
+/obj/structure/table/attack_hulk(mob/living/carbon/human/user)
+	..(user, 1)
+	visible_message("<span class='danger'>[user] smashes [src] apart!</span>")
+	playsound(src.loc, 'sound/effects/bang.ogg', 50, 1)
+	user.say(pick(";RAAAAAAAARGH!", ";HNNNNNNNNNGGGGGGH!", ";GWAAAAAAAARRRHHH!", "NNNNNNNNGGGGGGGGHH!", ";AAAAAAARRRGH!" ))
+	table_destroy(1)
+	return 1
+
 /obj/structure/table/attack_hand(mob/living/user)
 	user.changeNext_move(CLICK_CD_MELEE)
-	if(HULK in user.mutations)
-		user.do_attack_animation(src)
-		visible_message("<span class='danger'>[user] smashes [src] apart!</span>")
-		playsound(src.loc, 'sound/effects/bang.ogg', 50, 1)
-		user.say(pick(";RAAAAAAAARGH!", ";HNNNNNNNNNGGGGGGH!", ";GWAAAAAAAARRRHHH!", "NNNNNNNNGGGGGGGGHH!", ";AAAAAAARRRGH!" ))
-		table_destroy(1)
-		return
+	if(tableclimber && tableclimber != user)
+		tableclimber.Weaken(2)
+		tableclimber.visible_message("<span class='warning'>[tableclimber.name] has been knocked off the table", "You're knocked off the table!", "You see [tableclimber.name] get knocked off the table</span>")
+
 
 /obj/structure/table/attack_tk() // no telehulk sorry
 	return
@@ -241,8 +238,11 @@
 	else
 		return 0
 
-
-/obj/structure/table/MouseDrop_T(obj/O, mob/user)
+/obj/structure/table/MouseDrop_T(atom/movable/O, mob/user)
+	if(ismob(O) && user == O && ishuman(user))
+		if(user.canmove)
+			climb_table(user)
+			return
 	if ((!( istype(O, /obj/item/weapon) ) || user.get_active_hand() != O))
 		return
 	if(isrobot(user))
@@ -257,20 +257,23 @@
 	if(get_dist(src, user) < 2)
 		var/obj/item/weapon/grab/G = I
 		if(G.affecting.buckled)
-			user << "<span class='notice'>[G.affecting] is buckled to [G.affecting.buckled]!</span>"
-			return
+			user << "<span class='warning'>[G.affecting] is buckled to [G.affecting.buckled]!</span>"
+			return 0
 		if(G.state < GRAB_AGGRESSIVE)
-			user << "<span class='notice'>You need a better grip to do that!</span>"
-			return
+			user << "<span class='warning'>You need a better grip to do that!</span>"
+			return 0
 		if(!G.confirm())
-			return
+			return 0
 		G.affecting.loc = src.loc
 		G.affecting.Weaken(5)
 		G.affecting.visible_message("<span class='danger'>[G.assailant] pushes [G.affecting] onto [src].</span>", \
 									"<span class='userdanger'>[G.assailant] pushes [G.affecting] onto [src].</span>")
+		add_logs(G.assailant, G.affecting, "pushed")
+		qdel(I)
+		return 1
 	qdel(I)
 
-/obj/structure/table/attackby(obj/item/I, mob/user)
+/obj/structure/table/attackby(obj/item/I, mob/user, params)
 	if (istype(I, /obj/item/weapon/grab))
 		tablepush(I, user)
 		return
@@ -304,26 +307,23 @@
 			for(var/obj/item/C in oldContents)
 				C.loc = src.loc
 
-			user.visible_message("<span class='notice'>[user] empties [I] on [src].</span>")
+			user.visible_message("[user] empties [I] on [src].")
 			return
 		// If the tray IS empty, continue on (tray will be placed on the table like other items)
 
 	if(isrobot(user))
 		return
 
-	if(istype(I, /obj/item/weapon/melee/energy/blade))
-		var/datum/effect/effect/system/spark_spread/SS = new /datum/effect/effect/system/spark_spread()
-		SS.set_up(5, 0, src.loc)
-		SS.start()
-		playsound(src.loc, 'sound/weapons/blade1.ogg', 50, 1)
-		playsound(src.loc, "sparks", 50, 1)
-		user.visible_message("<span class='notice'>The [src.name] was sliced apart by [user]!</span>")
-		table_destroy(1)
-		return
-
 	if(!(I.flags & ABSTRACT)) //rip more parems rip in peace ;_;
 		if(user.drop_item())
 			I.Move(loc)
+			var/list/click_params = params2list(params)
+			//Center the icon where the user clicked.
+			if(!click_params || !click_params["icon-x"] || !click_params["icon-y"])
+				return
+			//Clamp it so that the icon never moves more than 16 pixels in either direction (thus leaving the table turf)
+			I.pixel_x = Clamp(text2num(click_params["icon-x"]) - 16, -(world.icon_size/2), world.icon_size/2)
+			I.pixel_y = Clamp(text2num(click_params["icon-y"]) - 16, -(world.icon_size/2), world.icon_size/2)
 
 
 /*
@@ -337,26 +337,31 @@
 /obj/structure/table/proc/table_destroy(var/destroy_type, var/mob/user)
 
 	if(destroy_type == TBL_DESTROY)
-		new framestack(src.loc)
-		new buildstack(src.loc)
+		for(var/i = 1, i <= framestackamount, i++)
+			new framestack(get_turf(src))
+		for(var/i = 1, i <= buildstackamount, i++)
+			new buildstack(get_turf(src))
 		qdel(src)
 		return
 
 	if(destroy_type == TBL_DISASSEMBLE)
-		user << "<span class='notice'>Now disassembling [src].</span>"
+		user << "<span class='notice'>You start disassembling [src]...</span>"
 		playsound(src.loc, 'sound/items/Screwdriver.ogg', 50, 1)
 		if(do_after(user, 20))
 			new frame(src.loc)
-			new buildstack(src.loc)
+			for(var/i = 1, i <= buildstackamount, i++)
+				new buildstack(get_turf(src))
 			qdel(src)
 			return
 
 	if(destroy_type == TBL_DECONSTRUCT)
-		user << "<span class='notice'>Now deconstructing [src].</span>"
+		user << "<span class='notice'>You start deconstructing [src]...</span>"
 		playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
 		if(do_after(user, 40))
-			new framestack(src.loc)
-			new buildstack(src.loc)
+			for(var/i = 1, i <= framestackamount, i++)
+				new framestack(get_turf(src))
+			for(var/i = 1, i <= buildstackamount, i++)
+				new buildstack(get_turf(src))
 			playsound(src.loc, 'sound/items/Deconstruct.ogg', 50, 1)
 			qdel(src)
 			return
@@ -365,23 +370,28 @@
  * TABLE CLIMBING
  */
 
-/obj/structure/table/MouseDrop_T(mob/target, mob/living/carbon/human/user)
-	if(istype(target) && user == target && istype(user))
-		if(user.canmove)
-			climb_table(user)
 
 /obj/structure/table/proc/climb_table(mob/user)
 	src.add_fingerprint(user)
 	user.visible_message("<span class='warning'>[user] starts climbing onto [src].</span>", \
-								"<span class='notice'>[user] starts climbing onto [src].</span>")
-	if(do_mob(user, user, 20))
-		user.pass_flags += PASSTABLE
-		step(user,get_dir(user,src.loc))
-		user.pass_flags -= PASSTABLE
-		user.visible_message("<span class='warning'>[user] climbs onto [src].</span>", \
-									"<span class='notice'>[user] climbs onto [src].</span>")
-		add_logs(user, src, "climbed onto")
-		user.Stun(2)
+								"<span class='notice'>You start climbing onto [src]...</span>")
+	var/climb_time = 20
+	if(user.restrained()) //Table climbing takes twice as long when restrained.
+		climb_time *= 2
+	tableclimber = user
+	if(do_mob(user, user, climb_time))
+		if(src.loc) //Checking if table has been destroyed
+			user.pass_flags += PASSTABLE
+			step(user,get_dir(user,src.loc))
+			user.pass_flags -= PASSTABLE
+			user.visible_message("<span class='warning'>[user] climbs onto [src].</span>", \
+									"<span class='notice'>You climb onto [src].</span>")
+			add_logs(user, src, "climbed onto")
+			user.Stun(2)
+			tableclimber = null
+			return 1
+	tableclimber = null
+	return 0
 
 
 /*
@@ -394,37 +404,17 @@
 	buildstack = /obj/item/stack/sheet/glass
 
 /obj/structure/table/glass/tablepush(obj/item/I, mob/user)
-	if(get_dist(src, user) < 2)
-		var/obj/item/weapon/grab/G = I
-		if(G.affecting.buckled)
-			user << "<span class='notice'>[G.affecting] is buckled to [G.affecting.buckled]!</span>"
-			return
-		if(G.state < GRAB_AGGRESSIVE)
-			user << "<span class='notice'>You need a better grip to do that!</span>"
-			return
-		if(!G.confirm())
-			return
-		G.affecting.loc = src.loc
-		G.affecting.Weaken(5)
-		G.affecting.visible_message("<span class='danger'>[G.assailant] smashes [G.affecting] into [src].</span>", \
-									"<span class='userdanger'>[G.assailant] smashes [G.affecting] into [src].</span>")
+	if(..())
+		visible_message("<span class='warning'>[src] breaks!</span>")
 		playsound(src.loc, "shatter", 50, 1)
 		new frame(src.loc)
 		new /obj/item/weapon/shard(src.loc)
 		qdel(src)
-	qdel(I)
+
 
 /obj/structure/table/glass/climb_table(mob/user)
-	src.add_fingerprint(user)
-	user.visible_message("<span class='warning'>[user] starts climbing onto [src].</span>", \
-								"<span class='notice'>[user] starts climbing onto [src].</span>")
-	if(do_mob(user, user, 20))
-		user.pass_flags += PASSTABLE
-		step(user,get_dir(user,src.loc))
-		user.pass_flags -= PASSTABLE
-		user.visible_message("<span class='warning'>[user] climbs onto [src], and it breaks!.</span>", \
-									"<span class='notice'>[user] climbs onto [src], and it breaks!.</span>")
-		add_logs(user, src, "climbed onto")
+	if(..())
+		visible_message("<span class='warning'>[src] breaks!</span>")
 		playsound(src.loc, "shatter", 50, 1)
 		new frame(src.loc)
 		new /obj/item/weapon/shard(src.loc)
@@ -459,23 +449,23 @@
 	var/status = 2
 	buildstack = /obj/item/stack/sheet/plasteel
 
-/obj/structure/table/reinforced/attackby(obj/item/weapon/W as obj, mob/user as mob)
+/obj/structure/table/reinforced/attackby(obj/item/weapon/W as obj, mob/user as mob, params)
 	if (istype(W, /obj/item/weapon/weldingtool))
 		var/obj/item/weapon/weldingtool/WT = W
 		if(WT.remove_fuel(0, user))
 			if(src.status == 2)
-				user << "<span class='notice'>Now weakening the reinforced table</span>"
+				user << "<span class='notice'>You start weakening the reinforced table...</span>"
 				playsound(src.loc, 'sound/items/Welder.ogg', 50, 1)
 				if (do_after(user, 50))
 					if(!src || !WT.isOn()) return
-					user << "<span class='notice'>Table weakened</span>"
+					user << "<span class='notice'>You weaken the table.</span>"
 					src.status = 1
 			else
-				user << "<span class='notice'>Now strengthening the reinforced table</span>"
+				user << "<span class='notice'>You start strengthening the reinforced table...</span>"
 				playsound(src.loc, 'sound/items/Welder.ogg', 50, 1)
 				if (do_after(user, 50))
 					if(!src || !WT.isOn()) return
-					user << "<span class='notice'>Table strengthened</span>"
+					user << "<span class='notice'>You strengthen the table.</span>"
 					src.status = 2
 			return
 	..()
@@ -483,19 +473,17 @@
 /obj/structure/table/reinforced/attack_paw(mob/user)
 	attack_hand(user)
 
-/obj/structure/table/reinforced/attack_hand(mob/user as mob)
-	user.changeNext_move(CLICK_CD_MELEE)
-	if ((HULK in user.mutations))
-		if (prob(75))
-			playsound(src, 'sound/effects/meteorimpact.ogg', 100, 1)
-			usr << text("<span class='notice'>You kick [src] into pieces.</span>")
-			usr.say(pick(";RAAAAAAAARGH!", ";HNNNNNNNNNGGGGGGH!", ";GWAAAAAAAARRRHHH!", "NNNNNNNNGGGGGGGGHH!", ";AAAAAAARRRGH!" ))
-			table_destroy(1)
-			return
-		else
-			playsound(src, 'sound/effects/bang.ogg', 50, 1)
-			usr << text("<span class='notice'>You kick [src].</span>")
-			return
+/obj/structure/table/reinforced/attack_hulk(mob/living/carbon/human/user)
+	..(user, 1)
+	if(prob(75))
+		playsound(src, 'sound/effects/meteorimpact.ogg', 100, 1)
+		user << text("<span class='notice'>You kick [src] into pieces.</span>")
+		user.say(pick(";RAAAAAAAARGH!", ";HNNNNNNNNNGGGGGGH!", ";GWAAAAAAAARRRHHH!", "NNNNNNNNGGGGGGGGHH!", ";AAAAAAARRRGH!" ))
+		table_destroy(1)
+	else
+		playsound(src, 'sound/effects/bang.ogg', 50, 1)
+		user << text("<span class='notice'>You kick [src].</span>")
+	return 1
 
 /*
  * Racks
@@ -510,26 +498,25 @@
 	throwpass = 1	//You can throw objects over this, despite it's density.
 	var/health = 5
 
-/obj/structure/rack/ex_act(severity)
+/obj/structure/rack/ex_act(severity, target)
 	switch(severity)
 		if(1.0)
 			qdel(src)
 		if(2.0)
-			qdel(src)
 			if(prob(50))
-				new /obj/item/weapon/rack_parts(src.loc)
+				rack_destroy()
+			else
+				qdel(src)
 		if(3.0)
 			if(prob(25))
-				qdel(src)
-				new /obj/item/weapon/rack_parts(src.loc)
+				rack_destroy()
 
 /obj/structure/rack/blob_act()
 	if(prob(75))
 		qdel(src)
 		return
 	else if(prob(50))
-		new /obj/item/weapon/rack_parts(src.loc)
-		qdel(src)
+		rack_destroy()
 		return
 
 /obj/structure/rack/CanPass(atom/movable/mover, turf/target, height=0)
@@ -547,23 +534,22 @@
 	if(isrobot(user))
 		return
 	if(!user.drop_item())
-		user << "<span class='notice'>\The [O] is stuck to your hand, you cannot put it in the rack!</span>"
+		user << "<span class='warning'>\The [O] is stuck to your hand, you cannot put it in the rack!</span>"
 		return
 	if (O.loc != src.loc)
 		step(O, get_dir(O, src))
 	return
 
-/obj/structure/rack/attackby(obj/item/weapon/W as obj, mob/user as mob)
+/obj/structure/rack/attackby(obj/item/weapon/W as obj, mob/user as mob, params)
 	if (istype(W, /obj/item/weapon/wrench))
-		new /obj/item/weapon/rack_parts( src.loc )
 		playsound(src.loc, 'sound/items/Ratchet.ogg', 50, 1)
-		qdel(src)
+		rack_destroy()
 		return
 
 	if(isrobot(user))
 		return
 	if(!user.drop_item())
-		user << "<span class='notice'>\The [W] is stuck to your hand, you cannot put it in the rack!</span>"
+		user << "<span class='warning'>\The [W] is stuck to your hand, you cannot put it in the rack!</span>"
 		return
 	W.Move(loc)
 	return 1
@@ -572,51 +558,63 @@
 /obj/structure/rack/attack_paw(mob/living/user)
 	attack_hand(user)
 
+/obj/structure/rack/attack_hulk(mob/living/carbon/human/user)
+	..(user, 1)
+	rack_destroy()
+	return 1
+
 /obj/structure/rack/attack_hand(mob/living/user)
 	user.changeNext_move(CLICK_CD_MELEE)
 	user.do_attack_animation(src)
 	playsound(loc, 'sound/items/dodgeball.ogg', 80, 1)
 	user.visible_message("<span class='warning'>[user] kicks [src].</span>", \
-						 "<span class='warning'>You kick [src].</span>")
-
-	if(HULK in user.mutations)
-		health -= 5
-	else
-		health -= rand(1,2)
+						 "<span class='danger'>You kick [src].</span>")
+	health -= rand(1,2)
 	healthcheck()
-
 
 /obj/structure/rack/attack_alien(mob/living/user)
 	user.do_attack_animation(src)
-	visible_message("<span class='danger'>[user] slices [src] apart!</span>")
-	new /obj/item/weapon/rack_parts(loc)
-	density = 0
-	qdel(src)
+	visible_message("<span class='warning'>[user] slices [src] apart.</span>")
+	rack_destroy()
 
 
 /obj/structure/rack/attack_animal(mob/living/simple_animal/user)
 	if(user.environment_smash)
 		user.do_attack_animation(src)
-		visible_message("<span class='danger'>[user] smashes [src] apart!</span>")
-		new /obj/item/weapon/rack_parts(loc)
-		density = 0
-		qdel(src)
+		visible_message("<span class='warning'>[user] smashes [src] apart.</span>")
+		rack_destroy()
 /obj/structure/rack/attack_tk() // no telehulk sorry
 	return
 
 /obj/structure/rack/proc/healthcheck()
 	if(health <= 0)
-		density = 0
-		var/obj/item/weapon/rack_parts/newparts = new(loc)
-		transfer_fingerprints_to(newparts)
-		qdel(src)
+		rack_destroy()
 	return
+
+/*
+ * Rack destruction
+ */
+
+/obj/structure/rack/proc/rack_destroy()
+	density = 0
+	var/obj/item/weapon/rack_parts/newparts = new(loc)
+	transfer_fingerprints_to(newparts)
+	qdel(src)
+
 
 /*
  * Rack Parts
  */
 
-/obj/item/weapon/rack_parts/attackby(obj/item/weapon/W as obj, mob/user as mob)
+/obj/item/weapon/rack_parts
+	name = "rack parts"
+	desc = "Parts of a rack."
+	icon = 'icons/obj/items.dmi'
+	icon_state = "rack_parts"
+	flags = CONDUCT
+	m_amt = 3750
+
+/obj/item/weapon/rack_parts/attackby(obj/item/weapon/W as obj, mob/user as mob, params)
 	..()
 	if (istype(W, /obj/item/weapon/wrench))
 		new /obj/item/stack/sheet/metal( user.loc )
@@ -625,10 +623,11 @@
 	return
 
 /obj/item/weapon/rack_parts/attack_self(mob/user as mob)
-	user << "<span class='notice'>Constructing rack...</span>"
+	user << "<span class='notice'>You start constructing rack...</span>"
 	if (do_after(user, 50))
+		if(!user.drop_item())
+			return
 		var/obj/structure/rack/R = new /obj/structure/rack( user.loc )
 		R.add_fingerprint(user)
-		user.drop_item()
 		qdel(src)
 		return
